General purpose config for Wetfish Teeworlds servers
This is a config I made for the Wetfish teeworlds servers. Was a bit
confused about what port to use, since they're supposed to fit in a
category. This doesn't, since it can switch between gametypes via the
votes. If you change the port dynamically to fit with the new gametype,
the clients wouldn't know to connect on the new port.